feat(用户管理): 添加用户登录IP和时间的记录与更新
feat(日志管理): 重构日志查询接口为分页查询并支持多条件筛选

refactor(用户服务): 优化用户更新逻辑，仅当角色变更时更新角色关系

docs(日志API): 更新日志管理接口文档，新增分页查询和批量删除接口

+++ zsk-module/zsk-module-system/src/main/java/com/zsk/system/service/impl/SysUserServiceImpl.java
@@ -109,8 +109,10 @@ public class SysUserServiceImpl extends ServiceImpl<SysUserMapper, SysUser> impl
     @Transactional(rollbackFor = Exception.class)
     public boolean updateUser(SysUser user) {
         Long userId = user.getId();
-        userRoleMapper.delete(new LambdaQueryWrapper<SysUserRole>().eq(SysUserRole::getUserId, userId));
-        insertUserRole(user);
+        if (user.getRoleIds() != null && user.getRoleIds().length > 0) {
+            userRoleMapper.delete(new LambdaQueryWrapper<SysUserRole>().eq(SysUserRole::getUserId, userId));
+            insertUserRole(user);
+        }
         return updateById(user);
     }
